requirements.txt
asgiref==3.11.0
certifi==2025.11.12
charset-normalizer==3.4.4
dj-database-url==3.0.1
Django==5.2.9
django-anymail==14.0
django-cors-headers==4.9.0
django-jet-reboot==1.3.10
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dotenv==0.9.9
drf-yasg==1.21.11
gunicorn==23.0.0
idna==3.11
inflection==0.5.1
mysqlclient==2.2.7
packaging==25.0
pillow==12.0.0
psycopg2-binary==2.9.11
PyJWT==2.10.1
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
sqlparse==0.5.4
uritemplate==4.2.0
urllib3==2.6.2
whitenoise==6.11.0
slack-sdk
